feat(StringInputWithCharacterCount): make maxCount required and show characters left
