+++ studio/components/stringInputWithCharacterCount/stringInputWithCharacterCount.module.css
@@ -1,0 +1,13 @@
+.srOnly {
+  /* visually hidden input to still receive focus and support screen readers */
+  /* https://developer.mozilla.org/en-US/docs/Web/API/File_API/Using_files_from_web_applications#using_a_label_element_to_trigger_a_hidden_file_input_element */
+  position: absolute !important;
+  height: 1px;
+  width: 1px;
+  overflow: hidden;
+  clip: rect(1px, 1px, 1px, 1px);
+}
+
+.overLimit {
+  color: rgb(203, 38, 24) !important;
+}
