+++ studio/schemas/documents/companyLocation.ts
@@ -1,5 +1,5 @@
 import { defineField, defineType } from "sanity";
-import { StringInputWithCharacterCount } from "../../components/StringInputWithCharacterCount";
+import { StringInputWithCharacterCount } from "../../components/stringInputWithCharacterCount/StringInputWithCharacterCount";
 
 export const companyLocationID = "companyLocation";
 export const companyLocationNameID = "companyLocationName";
