+++ studio/schemas/objects/footerSection.ts
@@ -2,7 +2,7 @@ import { defineType, StringInputProps } from "sanity";
 import { linkID } from "./link";
 import { soMeLinksID } from "../documents/socialMediaProfiles";
 import { richText, richTextID } from "../fields/text";
-import { StringInputWithCharacterCount } from "../../components/StringInputWithCharacterCount";
+import { StringInputWithCharacterCount } from "../../components/stringInputWithCharacterCount/StringInputWithCharacterCount";
 
 export const footerSectionID = {
   main: "footerSection",
