+++ studio/schemas/objects/sections/logoSalad.ts
@@ -1,7 +1,7 @@
 import { defineField, StringInputProps } from "sanity";
 import image from "studio/schemas/fields/media";
 import { richText } from "studio/schemas/fields/text";
-import { StringInputWithCharacterCount } from "../../../components/StringInputWithCharacterCount";
+import { StringInputWithCharacterCount } from "../../../components/stringInputWithCharacterCount/StringInputWithCharacterCount";
 
 export const logoSaladID = "logoSalad";
