+++ studio/schemas/objects/testimony.ts
@@ -1,6 +1,6 @@
 import { richText, title } from "../fields/text";
 import image from "../fields/media";
-import { StringInputWithCharacterCount } from "../../components/StringInputWithCharacterCount";
+import { StringInputWithCharacterCount } from "../../components/stringInputWithCharacterCount/StringInputWithCharacterCount";
 import { StringInputProps, StringRule } from "sanity";
 
 export const testimony = {
